refact scripts.

+++ docker/build-to-single-file.sh
@@ -1,7 +1,14 @@
 #!/bin/bash -e
+# $1 is target runtime id;
+if [ -z "$1" ]
+then
+  echo "Please provide target runtime id."
+  exit 1
+fi
+echo "Target runtime id is $1"
 rm -rf dist
 # Build server api image
-dotnet publish -r linux-x64 -c Release \
+dotnet publish -r $1 -c Release \
   -p:PublishSingleFile=true \
   -p:PublishTrimmed=false \
   -p:SelfContained=true \
@@ -12,7 +19,12 @@ dotnet publish -r linux-x64 -c Release \
 # modify config file to run in stagging server;
 sed -i.bak "s/ref=\"ConsoleAppender\"/ref=\"RollingFileAppender\"/g" dist/config/log.config
 sed -i.bak "s/DEBUG/ERROR/g" dist/config/log.config
-# sed -i.bak "s/127\.0\.0\.1/postgis/g" dist/config/hibernate.config
+# $2 is database server.
+if [ ! -z "$2" ]
+then
+  echo "Change database server to $2 "
+  sed -i.bak "s/127\.0\.0\.1/$2/g" dist/config/hibernate.config
+fi
 rm dist/config/*.bak
 # Build client project;
 cd ../client
@@ -27,4 +39,3 @@ cd ../docker
 # Copy client files to wwwroot
 cp -r ../client/dist dist/wwwroot
 rm -rf dist/wwwroot/app-shared
-# Build docker image
